Moved get_password_reset_token_expiry_time from views.py to models.py
